py/requirements.txt
google-cloud-asset==2.2.0

